atualizando valore de macros para Double

+++ app/src/main/java/com/example/dietasapp/data/model/FoodModel.kt
@@ -5,8 +5,8 @@ import java.io.Serializable
 data class FoodModel(
     val title: String = "",
     var calorie: Int = 0,
-    var fat: Int = 0,
-    var protein: Int = 0,
-    var carbohydrate: Int = 0,
+    var fat: Double = 0.0,
+    var protein: Double = 0.0,
+    var carbohydrate: Double = 0.0,
     var id: String = ""
 ): Serializable
